% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\docType{class}
\name{CelltypeSpatial-class}
\alias{CelltypeSpatial-class}
\alias{CelltypeSpatial}
\title{The CelltypeSpatial Class}
\description{
The CelltypeSpatial object stores Tangram mapped CelltypeSpatial data.
}
\section{Slots}{

\describe{
\item{\code{exp_mat}}{A data frame contains the Tangram mapped CelltypeSpatial data.}
}}

