% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\docType{class}
\name{Coordinate-class}
\alias{Coordinate-class}
\alias{Coordinate}
\title{The Coordinate Class}
\description{
The Coordinate object stores Coordinate data of cornplot.
}
\section{Slots}{

\describe{
\item{\code{coordinate}}{A data frame contains the x, y, and spot name of geoseq.}

\item{\code{text}}{The location name of geoseq.}

\item{\code{text_height}}{The y axis of text on cornplot.}

\item{\code{tick_start_of_layers}}{The start site of y tick.}

\item{\code{x}}{The range of x in cornplot.}

\item{\code{y}}{The range of y in cornplot.}

\item{\code{title_height}}{The height of title in cornplot.}

\item{\code{legend_position}}{The legend position in cornplot.}

\item{\code{shift}}{The shift of y tick.}

\item{\code{size}}{The dot size of spot.}

\item{\code{text_size}}{The text size of cornplot.}
}}

