% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{corn_plot}
\alias{corn_plot}
\title{Basic Cornplot}
\usage{
corn_plot(
  coord,
  data,
  title = NULL,
  stage,
  cols = c("azure2", "green", "yellow1", "yellow", "#FF3333", "#FF0000"),
  shapes = 1,
  limits = NULL
)
}
\arguments{
\item{coord}{Coordinate object from spatial object.}

\item{data}{A vector of plot data on each spot.}

\item{title}{The title plot on the head of cornplot.}

\item{stage}{The support stage of cornplot.}

\item{cols}{The color of cornplot.}

\item{shapes}{The shape of spot (1: "circle" or 2: "gaussian").}

\item{limits}{NULL to use the default scale range, or a numeric vector of length two providing limits of the scale.}
}
\value{
A ggplot2 object
}
\description{
Plot information on Cornplot, five stage is support now (E6.5, E6.75, E6.75_rep, E7.0, E7.25, E7.5, E7.5_suo).
}
