% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R, R/visualization.R
\name{SmoothData.spatial}
\alias{SmoothData.spatial}
\alias{CornPlot}
\alias{CornPlot.spatial}
\title{CornPlot}
\usage{
\method{CornPlot}{spatial}(
  object,
  key,
  omics = "snRNA_spatial",
  slot_name = "exp_mat",
  sigmas = 0.5,
  grid_smooth = "col"
)

CornPlot(object, ...)

\method{CornPlot}{spatial}(
  object,
  key,
  omics = "snRNASpatial",
  slot_name = "exp_mat",
  shapes = "circle",
  return_data = F,
  smooth = F,
  sigmas = 0.5,
  title = NULL,
  grid_smooth = "col"
)
}
\arguments{
\item{object}{A spatial object.}

\item{key}{A gene or peak in our dataset.}

\item{omics}{The omics to plot: 'snRNASpatial', 'snATACSpatial', 'Celltype_spatial', 'GEOseq'.}

\item{slot_name}{The slot name of data.}

\item{sigmas}{parameter in gaussian kernel, default:0.5.}

\item{grid_smooth}{Set smooth method: 'all', 'col', 'row', default 'col'.}

\item{...}{Arguments passed to other methods.}

\item{shapes}{The shape of spot (1: "circle" or 2: "gaussian").}

\item{return_data}{Return cornplot data, default False.}

\item{smooth}{Smooth cornplot data, default False.}

\item{title}{The title plot on the head of cornplot.}
}
\value{
Returns a cornplot.
}
\description{
CornPlot of multi-omics on cornplot.
}
