% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{CornPlotKeysets}
\alias{CornPlotKeysets}
\alias{CornPlotKeysets.spatial}
\title{CornPlotKeysets}
\usage{
CornPlotKeysets(object, ...)

\method{CornPlotKeysets}{spatial}(
  object,
  key,
  omics = "snRNASpatial",
  slot_name = "exp_mat",
  shapes = "circle",
  return_data = F,
  smooth = F,
  sigmas = 0.5,
  title = NULL,
  grid_smooth = "col"
)
}
\arguments{
\item{object}{A spatial object.}

\item{...}{Arguments passed to other methods.}

\item{key}{A gene or peak in our dataset.}

\item{omics}{The omics to plot: 'snRNASpatial', 'snATACSpatial', 'Celltype_spatial', 'GEOseq'.}

\item{slot_name}{The slot name of data.}

\item{shapes}{The shape of spot (1: "circle" or 2: "gaussian").}

\item{return_data}{Return cornplot data, default False.}

\item{smooth}{Smooth cornplot data, default False.}

\item{sigmas}{parameter in gaussian kernel, default:0.5.}

\item{title}{The title plot on the head of cornplot.}

\item{grid_smooth}{Set smooth method: 'all', 'col', 'row', default 'col'.}
}
\value{
Returns a cornplot.
}
\description{
Keysets of CornPlot of multi-omics on cornplot.
}
