% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{CornPlotMultiStage}
\alias{CornPlotMultiStage}
\title{Plot Multi stage CornPlot}
\usage{
CornPlotMultiStage(
  object,
  key,
  omics = "snRNASpatial",
  title = NULL,
  shapes = 1,
  scale = F,
  slot_name = "exp_mat",
  return_data = F,
  smooth = F,
  sigmas = 0.5,
  grid_smooth = "col"
)
}
\arguments{
\item{object}{A list of spatial object.}

\item{key}{A gene or peak in our dataset.}

\item{omics}{The omics to plot: 'snRNASpatial', 'snATACSpatial', 'Celltype_spatial', 'GEOseq'.}

\item{title}{The title of cornplot, the default is gene sysmbol.}

\item{shapes}{The shape of spot (1: "circle" or 2: "gaussian").}

\item{scale}{Unify expression into the same range, default False}

\item{slot_name}{The slot name of data.}

\item{return_data}{Return cornplot data, default False.}

\item{smooth}{Smooth cornplot data, default False.}

\item{sigmas}{parameter in gaussian kernel, default:0.5.}

\item{grid_smooth}{Set smooth method: 'all', 'col', 'row', default 'col'.}
}
\description{
Plot Multi stage CornPlot.
}
