% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{CreateCelltypeSpatialObject}
\alias{CreateCelltypeSpatialObject}
\title{Create a CelltypeSpatial object}
\usage{
CreateCelltypeSpatialObject(exp_mat)
}
\arguments{
\item{exp_mat}{A data frame contains Tangram mapped CelltypeSpatial data.}
}
\value{
A CelltypeSpatial object
}
\description{
Create a CelltypeSpatial object contain Tangram mapped CelltypeSpatial data.
}
