% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{CreateCoordinateObject}
\alias{CreateCoordinateObject}
\title{Create a Coordinate object}
\usage{
CreateCoordinateObject(
  coordinate,
  text,
  text_height,
  tick_start_of_layers,
  x,
  y,
  title_height,
  legend_position,
  shift,
  size,
  text_size
)
}
\arguments{
\item{coordinate}{A data frame contains the x, y, and spot name of geoseq.}

\item{text}{The location name of geoseq.}

\item{text_height}{The y axis of text on cornplot.}

\item{tick_start_of_layers}{The start site of y tick.}

\item{x}{The range of x in cornplot.}

\item{y}{The range of y in cornplot.}

\item{title_height}{The height of title in cornplot.}

\item{legend_position}{The legend position in cornplot.}

\item{shift}{The shift of y tick.}

\item{size}{The dot size of spot.}

\item{text_size}{The text size of cornplot.}
}
\value{
A Coordinate object
}
\description{
Create a Coordinate object contain the basic cornplot information.
}
