% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{CreateGEOseqObject}
\alias{CreateGEOseqObject}
\title{Create a GEOseq object}
\usage{
CreateGEOseqObject(exp_mat)
}
\arguments{
\item{exp_mat}{A data frame contains the log normalized GEO-seq data.}
}
\value{
A GEOseq object
}
\description{
Create a GEOseq object contain the expression of GEO-seq data.
}
