% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{CreatesnATACSpatialObject}
\alias{CreatesnATACSpatialObject}
\title{Create a snATACSpatial object}
\usage{
CreatesnATACSpatialObject(exp_mat, exp_mat2)
}
\arguments{
\item{exp_mat}{A data frame contains Tangram mapped snATACSpatial data.}

\item{exp_mat2}{Extra snATACSpatial data.}
}
\value{
A snATACSpatial object
}
\description{
Create a snATACSpatial object contain Tangram mapped snATACSpatial data.
}
