% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{CreatesnRNASpatialObject}
\alias{CreatesnRNASpatialObject}
\title{Create a snRNASpatial object}
\usage{
CreatesnRNASpatialObject(exp_mat)
}
\arguments{
\item{exp_mat}{A data frame contains Tangram mapped snRNASpatial data.}
}
\value{
A snRNASpatial object
}
\description{
Create a snRNASpatial object contain Tangram mapped snRNASpatial data.
}
