% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{CreateSpatialObject}
\alias{CreateSpatialObject}
\title{Create a spatial object}
\usage{
CreateSpatialObject(
  Coordinate_object,
  GEOseq_object,
  snRNASpatial_object,
  snATACSpatial_object,
  CelltypeSpatial_object,
  stage
)
}
\arguments{
\item{Coordinate_object}{Coordinate object.}

\item{GEOseq_object}{GEOseq object.}

\item{snRNASpatial_object}{snRNA spatial object.}

\item{snATACSpatial_object}{snATAC spatial object.}

\item{CelltypeSpatial_object}{celltype spatial object}

\item{stage}{The stage of embryo data (E6.5-E7.5).}
}
\value{
A spatial object
}
\description{
Create a spatial object from Coordinate_object, GEOseq_object, snRNASpatial_object, snATACSpatial_object and CelltypeSpatial_object.
}
