% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\docType{class}
\name{GEOseq-class}
\alias{GEOseq-class}
\alias{GEOseq}
\title{The GEOseq Class}
\description{
The GEOseq object stores GEOseq data.
}
\section{Slots}{

\describe{
\item{\code{exp_mat}}{A data frame contains the log normalized GEO-seq data.}
}}

