% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{SmoothData}
\alias{SmoothData}
\title{SmoothData}
\usage{
SmoothData(object, ...)
}
\arguments{
\item{object}{A spatial object.}

\item{...}{Arguments passed to other methods.}
}
\value{
Returns a cornplot.
}
\description{
Smooth cornplot data.
}
