% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\docType{class}
\name{snATACSpatial-class}
\alias{snATACSpatial-class}
\alias{snATACSpatial}
\title{The snATACSpatial Class}
\description{
The snATACSpatial object stores Tangram mapped snATACSpatial data.
}
\section{Slots}{

\describe{
\item{\code{exp_mat}}{A data frame contains the Tangram mapped snATACSpatial data.}

\item{\code{exp_mat2}}{Extra snATACSpatial data.}
}}

