% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\docType{class}
\name{snRNASpatial-class}
\alias{snRNASpatial-class}
\alias{snRNASpatial}
\title{The snRNASpatial Class}
\description{
The snRNASpatial object stores Tangram mapped snRNASpatial data.
}
\section{Slots}{

\describe{
\item{\code{exp_mat}}{A data frame contains the Tangram mapped snRNASpatial data.}
}}

