% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\docType{class}
\name{spatial-class}
\alias{spatial-class}
\alias{spatial}
\title{The spatial Class}
\description{
The spatial object is a wrapper of multi-omic gastrulation data.
}
\section{Slots}{

\describe{
\item{\code{Coordinate}}{Coordinate object.}

\item{\code{GEOseq}}{GEOseq object.}

\item{\code{snRNASpatial}}{snRNA spatial object.}

\item{\code{snATACSpatial}}{snATAC spatial object.}

\item{\code{CelltypeSpatial}}{Celltype spatial object.}

\item{\code{stage}}{The stage of embryo data (E6.5-E7.5).}
}}

