% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{SpatialFeature}
\alias{SpatialFeature}
\alias{SpatialFeature.spatial}
\title{SpatialFeature}
\usage{
SpatialFeature(object, ...)

\method{SpatialFeature}{spatial}(
  object,
  key,
  omics = "snRNA_spatial",
  slot_name = "exp_mat",
  top_feature = 100
)
}
\arguments{
\item{object}{A spatial object.}

\item{...}{Arguments passed to other methods.}

\item{key}{A vector of spot name.}

\item{omics}{The omics to plot: snRNA_spatial, snATAC_spatial, Celltype_spatial, GEOseq.}

\item{slot_name}{The slot name of data.}

\item{top_feature}{Number of top ranked features returned in the result.}
}
\value{
Returns a feature list.
}
\description{
Spatial Feature of multi-omics on cornplot.
}
